Reorganizes text parsing modules
* Creates an `annotations` module and relocates the
  relevant parser functions to it.
* Moves `top_level_stream_item` unit tests from the
  `reader` module to the `stream_item` module.
